Fix PR 36606 review issues
