+++ python/sglang/multimodal_gen/configs/transformers/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Transformers config compatibility classes for multimodal generation."""
