+++ python/sglang/multimodal_gen/runtime/parser/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Prompt and conversation parsing helpers for multimodal generation."""
