+++ python/sglang/multimodal_gen/runtime/pipelines/sensenova_u1.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 from typing import Any
 
 import torch
+from transformers import AutoModel, AutoTokenizer
 
 from sglang.multimodal_gen.configs.pipeline_configs.sensenova_u1 import (
     SenseNovaU1PipelineConfig,
@@ -12,8 +13,9 @@ from sglang.multimodal_gen.configs.sample.sensenova_u1 import (
     SenseNovaU1SamplingParams,
 )
 from sglang.multimodal_gen.runtime.disaggregation.roles import RoleType
-from sglang.multimodal_gen.runtime.models.sensenova_u1.loader import (
-    load_model_and_tokenizer,
+from sglang.multimodal_gen.runtime.distributed import get_local_torch_device
+from sglang.multimodal_gen.runtime.models import (  # noqa: F401
+    sensenova_u1 as _sensenova_u1,
 )
 from sglang.multimodal_gen.runtime.pipelines_core.composed_pipeline_base import (
     ComposedPipelineBase,
@@ -22,8 +24,10 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages import InputValidationS
 from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.sensenova_u1 import (
     SenseNovaU1GenerationStage,
 )
+from sglang.multimodal_gen.runtime.platforms import current_platform
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
+from sglang.multimodal_gen.runtime.utils.precision_types import PRECISION_TO_TYPE
 
 logger = init_logger(__name__)
 
@@ -54,7 +58,23 @@ class SenseNovaU1Pipeline(ComposedPipelineBase):
                 "SenseNovaU1Pipeline currently supports num_gpus=1. "
                 "Native tensor/pipeline parallelism is not implemented yet."
             )
-        modules = load_model_and_tokenizer(self.model_path, server_args)
+        dtype = PRECISION_TO_TYPE.get(
+            server_args.pipeline_config.model_precision, torch.bfloat16
+        )
+        model_kwargs: dict[str, Any] = {"torch_dtype": dtype}
+        tokenizer_kwargs: dict[str, Any] = {}
+        if server_args.trust_remote_code:
+            model_kwargs["trust_remote_code"] = True
+            tokenizer_kwargs["trust_remote_code"] = True
+        if server_args.revision is not None:
+            model_kwargs["revision"] = server_args.revision
+            tokenizer_kwargs["revision"] = server_args.revision
+
+        tokenizer = AutoTokenizer.from_pretrained(self.model_path, **tokenizer_kwargs)
+        model = AutoModel.from_pretrained(self.model_path, **model_kwargs).eval()
+        device = get_local_torch_device()
+        current_platform.set_device(device)
+        modules = {"model": model.to(device), "tokenizer": tokenizer}
         logger.info("Loaded SenseNova-U1 model from %s", self.model_path)
         return modules
