+++ python/sglang/multimodal_gen/test/unit/test_multi_output_grouping.py
@@ -6,7 +6,6 @@ import torch
 from sglang.multimodal_gen.configs.sample.sampling_params import SamplingParams
 from sglang.multimodal_gen.runtime.entrypoints.utils import (
     expand_request_outputs,
-    normalize_output_seeds,
 )
 from sglang.multimodal_gen.runtime.pipelines_core.schedule_batch import Req
 from sglang.multimodal_gen.runtime.pipelines_core.stages.base import PipelineStage
@@ -77,37 +76,56 @@ class CountingLatentStage(LatentPreparationStage):
 
 
 class TestMultiOutputGrouping(unittest.TestCase):
-    def test_normalize_output_seeds_from_int(self):
-        self.assertEqual(
-            normalize_output_seeds(10, num_outputs_per_prompt=3),
-            [10, 11, 12],
-        )
-
-    def test_normalize_output_seeds_from_per_prompt_list(self):
-        self.assertEqual(
-            normalize_output_seeds([3, 5], num_outputs_per_prompt=2),
-            [3, 5],
-        )
-
-    def test_normalize_output_seeds_from_total_list(self):
-        self.assertEqual(
-            normalize_output_seeds(
-                [1, 2, 3, 4],
-                num_outputs_per_prompt=2,
-                num_prompts=2,
-                prompt_index=1,
-            ),
-            [3, 4],
-        )
-
-    def test_normalize_output_seeds_rejects_mismatched_list(self):
-        with self.assertRaisesRegex(ValueError, r"seed list length"):
-            normalize_output_seeds(
-                [1, 2, 3],
-                num_outputs_per_prompt=2,
-                num_prompts=2,
-                prompt_index=0,
+    def test_expand_request_outputs_generates_seeds_from_int(self):
+        req = Req(
+            sampling_params=SamplingParams(
+                prompt="p",
+                num_outputs_per_prompt=3,
+                seed=10,
             )
+        )
+
+        outputs = expand_request_outputs(req)
+
+        self.assertEqual([item.seed for item in outputs], [10, 11, 12])
+
+    def test_expand_request_outputs_uses_per_prompt_seed_list(self):
+        req = Req(
+            sampling_params=SamplingParams(
+                prompt="p",
+                num_outputs_per_prompt=2,
+                seed=[3, 5],
+            )
+        )
+
+        outputs = expand_request_outputs(req)
+
+        self.assertEqual([item.seed for item in outputs], [3, 5])
+
+    def test_expand_request_outputs_slices_total_seed_list(self):
+        req = Req(
+            sampling_params=SamplingParams(
+                prompt="p",
+                num_outputs_per_prompt=2,
+                seed=[1, 2, 3, 4],
+            )
+        )
+
+        outputs = expand_request_outputs(req, num_prompts=2, prompt_index=1)
+
+        self.assertEqual([item.seed for item in outputs], [3, 4])
+
+    def test_expand_request_outputs_rejects_mismatched_seed_list(self):
+        req = Req(
+            sampling_params=SamplingParams(
+                prompt="p",
+                num_outputs_per_prompt=2,
+                seed=[1, 2, 3],
+            )
+        )
+
+        with self.assertRaisesRegex(ValueError, r"seed list length"):
+            expand_request_outputs(req, num_prompts=2, prompt_index=0)
 
     def test_expand_request_outputs_splits_seed_and_output_name(self):
         req = Req(
